# ai_connector/Github-Connector/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type LoginResultStep = typeof LoginResultStep[keyof typeof LoginResultStep];


export const LoginResultStep = {
  totp_setup: 'totp_setup',
  totp_verify: 'totp_verify',
} as const;

export interface LoginResult {
  step: LoginResultStep;
  message?: string;
}

export interface TotpSetup {
  secret: string;
  qrCodeUrl: string;
  otpauthUrl: string;
}

export interface TotpVerifyInput {
  token: string;
}

export interface AuthSession {
  authenticated: boolean;
  message?: string;
}

export interface AuthStatusResult {
  authenticated: boolean;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  role?: string | null;
}

export interface AccessRequestInput {
  email: string;
  name: string;
  /** @nullable */
  reason?: string | null;
}

export interface AccessRequestItem {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  reason?: string | null;
  status: string;
  createdAt: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface ChatMessage {
  id: number;
  sessionId: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatAttachment {
  name: string;
  type: string;
  data: string;
  size?: number;
}

export interface ChatMessageInput {
  sessionId: number;
  content: string;
  /** @nullable */
  systemPrompt?: string | null;
  attachments?: ChatAttachment[];
}

export interface ChatSession {
  id: number;
  title: string;
  createdAt: string;
  messageCount: number;
}

export interface ChatSessionInput {
  title: string;
}

export interface GithubRepo {
  id: number;
  name: string;
  fullName: string;
  /** @nullable */
  description?: string | null;
  url: string;
  private: boolean;
  /** @nullable */
  language?: string | null;
  updatedAt: string;
}

export interface GithubConfig {
  configured: boolean;
  /** @nullable */
  username?: string | null;
}

export interface GithubConfigInput {
  token: string;
}

export type ImageGenerateInputSize = typeof ImageGenerateInputSize[keyof typeof ImageGenerateInputSize];


export const ImageGenerateInputSize = {
  '1024x1024': '1024x1024',
  '1792x1024': '1792x1024',
  '1024x1792': '1024x1792',
} as const;

export interface ImageGenerateInput {
  prompt: string;
  size?: ImageGenerateInputSize;
}

export interface ImageGenerateResult {
  /** @nullable */
  url?: string | null;
  /** @nullable */
  b64?: string | null;
  /** @nullable */
  error?: string | null;
}

export type GetChatMessagesParams = {
sessionId?: string;
};