# ai_connector/Github-Connector/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
...
  /auth/verify-totp:
    post:
      operationId: verifyTotp
      tags: [auth]
      summary: Verify TOTP token and complete login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpVerifyInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid token
...
  /auth/request-access:
    post:
      operationId: requestAccess
      tags: [auth]
      summary: Request access to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessRequestInput"
      responses:
        "200":
          description: Request submitted
...
components:
  schemas:
...
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResult:
      type: object
      properties:
        step:
          type: string
          enum: [totp_setup, totp_verify]
        message:
          type: string
      required: [step]
...
    TotpVerifyInput:
      type: object
      properties:
        token:
          type: string
      required: [token]
    AuthSession:
      type: object
      properties:
        authenticated:
          type: boolean
        message:
          type: string
      required: [authenticated]
...
    AccessRequestInput:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        reason:
          type: ["string", "null"]
      required: [email, name]
...